Add files via upload
This page provides information about corruption, a significant issue with global political, economic, and environmental repercussions. It also includes factual statistics that are currently accessible. Given that corruption is a phenomenon with several facets, it is challenging to provide an accurate and thorough description. The premise that corrupt conduct involves the misuse of entrusted power for personal benefit, however, is at the heart of the majority of definitions of corruption. Bribery, clientelism, and embezzlement are classic instances. 